requirements.txt
# Deep Learning Framework (CUDA 11.8)
# Note: Use latest version for local development
# Competition environment (AI Factory) will use specific versions in task.ipynb
torch==2.7.1+cu118
torchvision==0.22.1+cu118

# Model Architecture
timm==0.9.12

# Computer Vision & Video Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Data Augmentation
albumentations==1.3.1

# Face Detection
facenet-pytorch==2.5.3
mediapipe==0.10.21

# Scientific Computing
# numpy is installed by PyTorch
scipy>=1.11.4

# Machine Learning
scikit-learn>=1.3.2
pandas>=2.0.3

# Configuration Management
pyyaml==6.0.1

# Progress Bars
tqdm==4.66.1

# Visualization (for EDA notebooks)
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest==7.4.3

# Competition Submission
aifactory
